requirements.txt
rpi_ws281x
adafruit-circuitpython-neopixel
LD2410
pyserial
pug4py
aiohttp
python-decouple